% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisar_radar_PR.R
\name{analisar_radar_PR}
\alias{analisar_radar_PR}
\title{Analisa a cor média em uma região circular do radar e interpreta intensidade de chuva}
\usage{
analisar_radar_PR(img, mega = "Castelo", raio)
}
\arguments{
\item{img}{Imagem de radar obtida pela função \code{baixar_radar_PR()}.}

\item{mega}{Nome da cidade/ponto de interesse. Aceita: "Cianorte", "Castelo", "teste".}

\item{raio}{Raio em pixels da área circular a ser analisada (default = 55).}
}
\value{
Texto indicando a condição meteorológica: "Chuva forte", "Chuva leve", "Risco de chuva" ou "Sem chuvas".
}
\description{
Esta função analisa uma imagem de radar meteorológico e classifica a intensidade de chuva com base na cor (RGB) de uma área circular ao redor de coordenadas específicas.
}
\examples{
\dontrun{
img <- baixar_radar_PR()
result <- analisar_radar_PR(img, mega = "Cianorte", raio = 55)
}

}
\author{
Santos Henrique Brant Dias
}
