% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_radar_PR.R
\name{baixar_radar_PR}
\alias{baixar_radar_PR}
\title{Faz o download de imagem de radar meteorológico do PR (Simepar)}
\usage{
baixar_radar_PR()
}
\value{
Uma imagem (magick-image) contendo o radar meteorológico atual do estado do Paraná.
}
\description{
Esta função acessa a imagem mais recente do radar do Paraná disponibilizada pelo Simepar e retorna uma imagem tipo magick.
}
\examples{
\dontrun{
radar_img <- baixar_radar_PR()
}

}
\author{
Santos Henrique Brant Dias
}
