% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_curvas_nivel.R
\name{criar_curvas_nivel}
\alias{criar_curvas_nivel}
\title{Curvas de nível com buffer automático (em metros) e recorte pelo polígono}
\usage{
criar_curvas_nivel(
  raster_altitude,
  kml,
  intervalo = 1,
  buffer_dist = 200,
  ajust = T
)
}
\arguments{
\item{raster_altitude}{Raster de entrada (\code{SpatRaster}) ou caminho para arquivo raster (.tif)}

\item{kml}{Objeto \code{sf} ou caminho/URL para arquivo KML com a área de interesse}

\item{intervalo}{Intervalo das curvas de nível em metros (padrão = 5)}

\item{buffer_dist}{Distância do buffer em metros (padrão = 200)}

\item{ajust}{Para fazer o corte usando o kml para sair identico ao kml se for T}
}
\value{
Objeto \code{sf} com as curvas de nível e atributo \code{elev}
}
\description{
Esta função gera curvas de nível (contornos) de um raster de altitude.
Aplica buffer em torno da área de interesse, corta e mascara o raster,
e retorna um objeto \code{sf} com as curvas de nível no CRS WGS84 UTM.
}
\details{
Gera curvas de nível a partir de um raster de altitude e um polígono (KML/SF)
}
\note{
\itemize{
\item Para rasters muito grandes, considere recortar previamente para melhorar desempenho.
\item Se o raster tiver valores \code{NA} extensos na área, alguns níveis podem não gerar linhas.
}
}
\section{Requisitos}{

Pacotes: \code{sf}, \code{terra}.
}

\examples{
 \dontrun{
curvas <- criar_curvas_nivel(raster_altitude="dem_sad69.tif", kml="area.kml", intervalo=10, buffer_dist=500, ajust = T)
}

}
\seealso{
\code{\link[terra:as.contour]{terra::as.contour()}}, \code{\link[sf:st_buffer]{sf::st_buffer()}}, \code{\link[terra:mask]{terra::mask()}}
}
\author{
Santos Henrique Brant Dias
}
