% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eto_fao56_station.R
\name{eto_fao56_station}
\alias{eto_fao56_station}
\title{Calcula a Evapotranspiração de Referência diária (ETo) pelo método FAO 56}
\usage{
eto_fao56_station(df, z, lat)
}
\arguments{
\item{df}{Dataframe com dados horários meteorológicos.}

\item{z}{Altitude da estação (em metros) — pode ser um valor único ou vetor nomeado por estação.}

\item{lat}{Latitude da estação em graus decimais — pode ser um valor único ou vetor nomeado.}
}
\value{
Um data.frame com a ETo diária por estação e data.
}
\description{
A função calcula a ETo diária agrupada por estação e data, com base no método de Penman-Monteith da FAO.
}
