% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clima_estacao.R
\name{plot_clima_estacao}
\alias{plot_clima_estacao}
\title{Plota gráficos climáticos (Temperatura, Ponto de Orvalho, Umidade e etc)}
\usage{
plot_clima_estacao(df, estacao, datas)
}
\arguments{
\item{df}{Dataframe com os dados meteorológicos processados.}

\item{estacao}{Nome da estação desejada (ex: "ICIANO1").}

\item{datas}{Vetor de datas a serem plotadas (formato: "YYYY-MM-DD").}
}
\value{
Vários gráficos plotados diretamente no dispositivo gráfico do R.
}
\description{
Esta função plota gráficos de temperatura do ar, ponto de orvalho e umidade relativa do ar ao longo do tempo,
utilizando dados meteorológicos processados por \code{station_wund_download()}. Permite filtrar por estação e intervalo de datas,
e apresenta a umidade relativa em eixo secundário.
}
\examples{
\dontrun{
plot_clima_estacao(df,
                   estacao = "IPARANAM3",
                   datas = c("2025-06-11","2025-06-12"))
}

}
\author{
Santos Henrique Brant Dias
}
