% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_to_random_points.R
\name{polygon_to_random_points}
\alias{polygon_to_random_points}
\title{Está função serve para criar pontos a partir de de um poligono}
\usage{
polygon_to_random_points(
  polygon,
  Npoints,
  min_dist,
  borda = 20,
  pt = TRUE,
  N = TRUE
)
}
\arguments{
\item{polygon}{Caminho do arquivo do polígono vetorial}

\item{Npoints}{Número de pontos para plotar}

\item{min_dist}{Valor da distancia entre os pontos em metros}

\item{borda}{Distancia da borda para não colocar pontos}

\item{pt}{True ou FALSE para ver o plot do arquivo}

\item{N}{True ou FALSE para}
}
\value{
Returns um arquivo vetorial (ex. KML)
}
\description{
Função gera um grade pontos regulares para realização de análises, informações geradas a partir arquivo vetorial tipo polígono, que pode ser feito pelo Google Esrth Pro com um arquivo KML.
}
\examples{
if (interactive()) {
polygon_to_random_points(dir_polygon = "./Downloads/Demilitacao_Area.kml",
dist = 100, pt = TRUE)
}

}
\author{
Santos Henrique Brant Dias
}
