% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoData_download_to_vector.R
\name{TopoData_download_to_vector}
\alias{TopoData_download_to_vector}
\title{Está função serve para fazer o download de dados de estações da wunderground}
\usage{
TopoData_download_to_vector(vector, layer = "Declividade")
}
\arguments{
\item{vector}{Caminho do arquivo do polígono vetorial}

\item{layer}{Valor da distancia entre os pontos em metros}
}
\value{
Raster do TopoData
}
\description{
Função faz o download do TOPODATA INPE. Pode ser feito o download das seguintes variavéis: Altitude, Declividade, RelevoSombreado, Orientação, FormaTerreno, DivisoresTalvegues, Curv.Vertical, Curv.Horizontal
}
\examples{
if (interactive()) {
Altitude <- TopoData_download_to_vector(vector = "C:/User/Downloads/area.kml", layer = "Altitude")
}

}
\author{
Santos Henrique Brant Dias
}
