% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eto_fao56_station.R
\docType{data}
\name{z}
\alias{z}
\title{Calcula a Evapotranspiração de Referência diária (ETo) pelo método FAO 56}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
z
}
\arguments{
\item{df}{Dataframe com dados meteorológicos horários, contendo ao menos as colunas:
\code{Estacao}, \code{DataHora}, \code{Temperatura_C}, \code{PontoOrvalho_C}, \code{RadiacaoSolar_W_m2},
\code{VelocidadeVento_m_s} e \code{Pressao_hPa}.}

\item{z}{Altitude da estação em metros. Pode ser um valor único (para todas as estações) ou
um vetor nomeado com os nomes das estações como \code{c("EST1" = 650, "EST2" = 780)}.}

\item{lat}{Latitude da estação em graus decimais. Pode ser um valor único ou um vetor nomeado, como em \code{z}.}
}
\value{
Um data.frame contendo a ETo diária por estação e por data, com colunas: \code{Estacao}, \code{Data}, \code{ETo_mm_dia}.
}
\description{
Esta função calcula a evapotranspiração de referência diária (ETo) utilizando o método de Penman-Monteith proposto pela FAO (FAO-56).
Os dados de entrada devem conter observações meteorológicas horárias incluindo temperatura do ar, ponto de orvalho, radiação solar,
velocidade do vento e pressão atmosférica. A função agrupa os dados por estação e por dia e retorna a ETo diária.

A altitude (z) e a latitude (lat) da estação devem ser fornecidas como valores únicos ou vetores nomeados com os nomes das estações.
}
\details{
A equação de Penman-Monteith requer dados médios diários de temperatura, pressão atmosférica,
déficit de vapor de pressão, radiação líquida e resistência aerodinâmica. A função realiza as conversões
necessárias a partir de dados horários.
}
\examples{
\dontrun{
eto_df <- calcular_eto_fao56(df, z = 600, lat = -23.4)
}

}
\author{
Santos Henrique Brant Dias
}
\keyword{datasets}
