% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eto_fao56_station.R
\name{eto_fao56_station}
\alias{eto_fao56_station}
\title{Calcula a Evapotranspiração de Referência diária (ETo) pelo método FAO 56}
\usage{
eto_fao56_station(df, estacao)
}
\arguments{
\item{df}{Dataframe com dados meteorológicos horários.}

\item{estacao}{Nome da estação desejada (ex: "ICIANO1").}
}
\value{
Um data.frame contendo a ETo diária por estação e por data.
}
\description{
Esta função calcula a evapotranspiração de referência diária (ETo) utilizando o método de Penman-Monteith proposto pela FAO (FAO-56).
Os dados de entrada devem conter observações meteorológicas horárias incluindo temperatura do ar, ponto de orvalho, radiação solar,
velocidade do vento e pressão atmosférica. A função agrupa os dados por estação e por dia e retorna a ETo diária.

A altitude (z) e a latitude (lat) da estação devem ser fornecidas como valores únicos ou vetores nomeados com os nomes das estações.
}
\details{
A equação de Penman-Monteith requer dados médios diários de temperatura, pressão atmosférica,
déficit de vapor de pressão, radiação líquida e resistência aerodinâmica. A função realiza as conversões
necessárias a partir de dados horários.
}
\examples{
\dontrun{
eto_df <- eto_fao56_station(df, estacao='ICIANO1')
}

}
\author{
Santos Henrique Brant Dias
}
