% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executar_alerta_telegram.R
\name{executar_alerta_telegram}
\alias{executar_alerta_telegram}
\title{Executa análise do radar meteorológico e envia alerta com imagem via Telegram}
\usage{
executar_alerta_telegram(
  mega = "Cianorte",
  img_plot,
  chat_id,
  bot_token,
  raio = 50
)
}
\arguments{
\item{mega}{Nome da cidade a ser monitorada. Deve ser um dos nomes previamente cadastrados na função (ex: "Cianorte", "Castelo").}

\item{chat_id}{Código identificador do grupo ou chat do Telegram (ex: "-1001234567890").}

\item{bot_token}{Token do bot do Telegram (obtido via @BotFather).}
}
\value{
Não retorna valor, mas envia uma mensagem automática ao Telegram em caso de detecção de chuva.
}
\description{
Esta função realiza a análise da imagem de radar meteorológico do estado do Paraná (via Simepar),
identificando a ocorrência de chuva nas proximidades de uma cidade especificada. Em caso positivo, uma imagem
com marcações geográficas e legenda é enviada automaticamente para um grupo ou chat no Telegram.
}
\details{
A imagem de radar é obtida em tempo real do Simepar. A função compara os valores RGB de pixels centrais da cidade para inferir a presença de chuva.
A mensagem enviada inclui imagem com marcações e legenda no formato Markdown.
}
\author{
Santos Henrique Brant Dias
}
