% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_imagem_radar.R
\name{gerar_imagem_radar}
\alias{gerar_imagem_radar}
\title{Gera imagem do radar meteorológico com marcações circulares e retorna imagem editada}
\usage{
gerar_imagem_radar(cidade, raio)
}
\arguments{
\item{raio}{Raio da área de análise em pixels (default: 40)}

\item{coords}{Lista com coordenadas nomeadas das cidades. Exemplo: list("Cianorte" = list(x=388, y=240))}
}
\value{
Objeto de imagem com as marcações, ou NULL se houver falha no download
}
\description{
Esta função baixa a imagem do radar meteorológico do Paraná e adiciona marcações circulares
ao redor das coordenadas de cidades de interesse. A imagem gerada pode ser usada para envio
por Telegram ou outras finalidades.
}
\examples{

coords <- list(
  'Cianorte' = list(x = 388, y = 240),
  'Castelo'  = list(x = 437, y = 190)
)
img_plot <- gerar_imagem_radar(coords)

}
\author{
Santos Henrique Brant Dias
}
