% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_wund_download.R
\name{station_wund_download}
\alias{station_wund_download}
\title{Está função serve para fazer o download de dados de estações da wunderground}
\usage{
station_wund_download(stations, start_date, end_date)
}
\arguments{
\item{stations}{Vetor com um ou mais códigos de estação (ex: "ICIANO1").}

\item{start_date}{Data inicial no formato "YYYY-MM-DD".}

\item{end_date}{Data final no formato "YYYY-MM-DD".}
}
\value{
Um data.frame com os dados meteorológicos.
}
\description{
Função faz o download de dados meteorológicos wunderground.
}
\examples{
\dontrun{
station_wund_download(stations = c("ICIANO1", "IMANDA28"),
                      start_date = "2024-12-01",
                      end_date = "2024-12-03")
}

}
\author{
Santos Henrique Brant Dias
}
