% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoData_download_to_vector.R
\name{TopoData_download_to_vector}
\alias{TopoData_download_to_vector}
\title{Está função serve para fazer o download de dados de estações da wunderground}
\usage{
TopoData_download_to_vector(vector, pathdow)
}
\arguments{
\item{vector}{Caminho do arquivo do polígono vetorial}

\item{pathdow}{Valor da distancia entre os pontos em metros}
}
\value{
Raster do TopoData
}
\description{
Função faz o download de dados meteorológicos wunderground.
}
\examples{
TopoData_download_to_vector("C:/Users/server_SantosDias/Downloads/Demilitacao_Area.kml",
"C:/Users/server_SantosDias/Downloads")


}
\author{
Santos Henrique Brant Dias
}
